# Improve UX of initial sync logs, pushing not relevant logs to debug l…