--- a/ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/EngineForkchoiceUpdated.java
+++ b/ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/EngineForkchoiceUpdated.java
@@ -38,6 +38,7 @@ import org.hyperledger.besu.ethereum.api.jsonrpc.internal.results.EngineUpdateFo
 import org.hyperledger.besu.ethereum.core.BlockHeader;
 
 import java.util.Optional;
+import java.util.function.BiConsumer;
 
 import io.vertx.core.Vertx;
 import org.slf4j.Logger;
@@ -265,25 +266,49 @@ public class EngineForkchoiceUpdated extends ExecutionEngineJsonRpcMethod {
   private JsonRpcResponse syncingResponse(
       final Object requestId, final EngineForkchoiceUpdatedParameter forkChoice) {
 
-    logForkchoiceUpdatedCall(SYNCING, forkChoice);
+    logForkchoiceUpdatedCall(this::logAtDebug, SYNCING, forkChoice);
     return new JsonRpcSuccessResponse(
         requestId, new EngineUpdateForkchoiceResult(SYNCING, null, null, Optional.empty()));
   }
 
   // fcU calls are synchronous, no need to make volatile
   private long lastFcuInfoLog = System.currentTimeMillis();
+  private static final String logMessage =
+      "{} for fork-choice-update: head: {}, finalized: {}, safeBlockHash: {}";
 
   private void logForkchoiceUpdatedCall(
       final EngineStatus status, final EngineForkchoiceUpdatedParameter forkChoice) {
+    logForkchoiceUpdatedCall(this::logAtInfo, status, forkChoice);
+  }
+
+  private void logForkchoiceUpdatedCall(
+      final BiConsumer<EngineStatus, EngineForkchoiceUpdatedParameter> logAtLevel,
+      final EngineStatus status,
+      final EngineForkchoiceUpdatedParameter forkChoice) {
     // cheaply limit the noise of fcU during consensus client syncing to once a minute:
     if (lastFcuInfoLog + ENGINE_API_LOGGING_THRESHOLD < System.currentTimeMillis()) {
       lastFcuInfoLog = System.currentTimeMillis();
-      LOG.info(
-          "{} for fork-choice-update: head: {}, finalized: {}, safeBlockHash: {}",
-          status.name(),
-          forkChoice.getHeadBlockHash(),
-          forkChoice.getFinalizedBlockHash(),
-          forkChoice.getSafeBlockHash());
+      logAtLevel.accept(status, forkChoice);
     }
+  }
+
+  private void logAtInfo(
+      final EngineStatus status, final EngineForkchoiceUpdatedParameter forkChoice) {
+    LOG.info(
+        logMessage,
+        status.name(),
+        forkChoice.getHeadBlockHash(),
+        forkChoice.getFinalizedBlockHash(),
+        forkChoice.getSafeBlockHash());
+  }
+
+  private void logAtDebug(
+      final EngineStatus status, final EngineForkchoiceUpdatedParameter forkChoice) {
+    LOG.debug(
+        logMessage,
+        status.name(),
+        forkChoice.getHeadBlockHash(),
+        forkChoice.getFinalizedBlockHash(),
+        forkChoice.getSafeBlockHash());
   }
 }