@non-fungible-nelson 2022-10-06
Does this alter FCU updates if the CL is in sync while Besu is syncing? The other PR just changes the language of the FCU update from Besu SYNCING language for clarity.