@fab-10 2022-10-06
this does logs the FcUs while syncing at debug level, so they will not be visible by default, and restore them at info level when the sync is done. This because the FcUs are very frequent during the initial sync and they hide more relevant sync progress logs.
Do you think this is acceptable?